Create a customer and payment method if ID matches
Create a customer and payment method in the registry if the user ID
provided to Braintree::Customer.find matches the user ID specified in
the environment variable `TEST_USER_ID`.

This is to help with load tests that expect the default test user to
have a payment method saved in Braintree (e.g. when purchasing products
on the store).